requirements.txt
# HAI ICS Security Pipeline — Requirements
# Core data science
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0
scikit-learn>=1.2.0

# Gradient boosting models
xgboost>=1.7.0
lightgbm>=3.3.0

# Dashboard
streamlit>=1.28.0
plotly>=5.14.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Utilities
pyyaml>=6.0
joblib>=1.2.0
tqdm>=4.65.0

# Deep learning (optional — comment out if deploying to free Streamlit Cloud)
# torch>=2.0.0

# Explainability (optional)
# shap>=0.41.0
